Refactor tests to extract common logic
