+++ loader/src/test/java/net/neoforged/fml/loading/MixinConfigTest.java
@@ -8,90 +8,16 @@ package net.neoforged.fml.loading;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import net.neoforged.fml.ModLoadingException;
-import net.neoforged.fml.loading.mixin.FMLMixinService;
 import net.neoforged.fml.loading.mixin.MixinFacade;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.spongepowered.asm.launch.MixinBootstrap;
-import org.spongepowered.asm.launch.platform.MixinPlatformManager;
-import org.spongepowered.asm.mixin.MixinEnvironment;
 import org.spongepowered.asm.mixin.Mixins;
-import org.spongepowered.asm.mixin.transformer.Config;
-import org.spongepowered.asm.service.MixinService;
-import org.spongepowered.asm.service.ServiceNotAvailableError;
 
-public class MixinConfigTest extends LauncherTest {
-    @BeforeEach
-    void cleanUpMixins() {
-        // What a mess. If we never successfully initialized Mixin
-        // we cannot even load the class, since it'd try to initialize a default service.
-        // See MixinFacade for details
-        if (System.getProperty("mixin.service") == null) {
-            return;
-        }
-
-        try {
-            Mixins.getConfigs().clear();
-        } catch (ServiceNotAvailableError ignored) {}
-
-        clearCollection(Config.class, null, "allConfigs");
-        clearCollection(Mixins.class, null, "registeredConfigs");
-
-        var service = (FMLMixinService) MixinService.getService();
-        service.clearMixinContainers();
-
-        // Clear root platform manager
-        var platform = MixinBootstrap.getPlatform();
-        clearCollection(MixinPlatformManager.class, platform, "containers");
-        setField(MixinPlatformManager.class, platform, "injected", false);
-        setField(MixinBootstrap.class, null, "platform", null); // Make MixinBootstrap call init on the platform again
-
-        gotoPhase(MixinEnvironment.Phase.PREINIT);
-    }
-
-    private void gotoPhase(MixinEnvironment.Phase phase) {
-        try {
-            var m = MixinEnvironment.class.getDeclaredMethod("gotoPhase", MixinEnvironment.Phase.class);
-            m.setAccessible(true);
-            m.invoke(null, phase);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private static <T> void setField(Class<?> clazz, T instance, String fieldName, Object value) {
-        try {
-            var field = clazz.getDeclaredField(fieldName);
-            field.setAccessible(true);
-            field.set(instance, value);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            throw new RuntimeException("Failed to set field value " + fieldName, e);
-        }
-    }
-
-    private static <T> void clearCollection(Class<T> clazz, T instance, String fieldName) {
-        Object obj;
-        try {
-            var field = clazz.getDeclaredField(fieldName);
-            field.setAccessible(true);
-            obj = field.get(instance);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            throw new RuntimeException("Failed to get field value " + fieldName, e);
-        }
-        switch (obj) {
-            case Map<?, ?> map -> map.clear();
-            case Collection<?> collection -> collection.clear();
-            case null -> {}
-            default -> throw new IllegalStateException("Don't know how to clear " + obj);
-        }
-    }
-
+public class MixinConfigTest extends LauncherTest implements MixinTestHelper {
     @Test
     void testRequiredModIsMissing() throws Exception {
         installation.setupProductionClient();
@@ -204,62 +130,6 @@ public class MixinConfigTest extends LauncherTest {
         launchAndLoad("neoforgeclient");
 
         assertThat(Mixins.getConfigs()).extracting("name").containsOnly("test.mixins.json");
-    }
-
-    @Test
-    void testFrameRecomputationWithGeneratedClasses() throws Exception {
-        installation.setupProductionClient();
-        installation.buildModJar("mixin-test.jar")
-                .withTestmodModsToml(modsToml -> modsToml.addMixinConfig("test.mixins.json").addMod("test"))
-                .addTextFile("test.mixins.json", """
-                        {
-                            "package": "test.mixin",
-                            "mixins": ["RequiresFrameRecompute"]
-                        }
-                        """)
-                .addClass("test.target.Super", """
-                        public class Super {}
-                        """)
-                .addClass("test.target.Target", """
-                        import net.neoforged.fml.common.Mod;
-
-                        @Mod("test")
-                        public class Target {
-                            {
-                                test(null);
-                            }
-
-                            public static void test(Super obj) {}
-                        }
-                        """)
-                .addClass("test.mixin.RequiresFrameRecompute", """
-                        import org.spongepowered.asm.mixin.injection.At;
-                        import org.spongepowered.asm.mixin.injection.Inject;
-                        import org.spongepowered.asm.mixin.Mixin;
-                        import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-                        import test.target.Super;
-
-                        @Mixin(test.target.Target.class)
-                        public class RequiresFrameRecompute {
-                            @Inject(
-                                method = "test",
-                                at = @At("HEAD")
-                            )
-                            private static void testInject(Super obj, CallbackInfo ci) {
-                                // makes recomputing the stack necessary
-                                takesLocalType(alwaysFalse() ? obj : new Super() {});
-                            }
-
-                            private static boolean alwaysFalse() {
-                                return false;
-                            }
-
-                            private static void takesLocalType(Super object) {}
-                        }
-                        """)
-                .build();
-
-        var result = launchAndLoad("neoforgeclient");
     }
 
     /**
